Linked List in C.
It links the data 0-9 in a linked list, and prints the data
on the console in the same way as it was enterd.

Liked list as a concept is easy. Implementing the same in C
required understanding of memory allocation and pointers.
I'd recommend reading the use of struct and pointers, and a decent
understanding of linkedlists before getting into the code.

commit 2

commit 3

commit 4